requirement.txt
Flask-SQLAlchemy==3.1.1
black==23.9.1
pynvim==0.4.3
flask-restx==1.2.0
